% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_daily.R
\name{all_daily}
\alias{all_daily}
\title{Combine all GGIR daily output into a single dataset}
\usage{
all_daily(filepath)
}
\arguments{
\item{filepath}{A filepath to the folder containing your results}
}
\description{
This function combines all 'daily' parts of the GGIR output (within a
single results folder) into a single dataset. So part2_daysummary,
part4_nightsummary_sleep_cleaned, and part5_daysummary will be combined by ID
number.
}
