% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_summary.R
\name{all_summary}
\alias{all_summary}
\title{Combine all GGIR summary output into a single dataset}
\usage{
all_summary(filepath)
}
\arguments{
\item{filepath}{A filepath to the folder containing your results}
}
\description{
This function combines all 'summary' parts of the GGIR output (within a
single results folder) into a single dataset. So part2_summary,
part4_summary_sleep_cleaned, and part5_personsummary will be combined by ID
number.
}
