% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_summary.R
\name{get_summary_parts}
\alias{get_summary_parts}
\title{Combine GGIR Summary Output}
\usage{
get_summary_parts(filepath, part)
}
\arguments{
\item{filepath}{A filepath to the folder holding all of your output data}

\item{part}{The output you want to combine, i.e. 2, 4 or 5}
}
\description{
This function allows you to combine multiple csv files from the GGIR
output, to get data from all your participants in the same dataset
}
